Create Tree:Level_Order_Traversal
Given a pointer to the root of a binary tree, you need to print the level order traversal of this tree. In level-order traversal, nodes are visited level by level from left to right. Complete the function levelOrder and print the values in a single line separated by a space.

For example:

     1
      \
       2
        \
         5
        /  \
       3    6
        \
         4  
For the above tree, the level order traversal is .
1->2->5->3->6->4
Input Format

You are given a function,

void levelOrder(Node * root) {

}
Constraints
1<=Nodes in the tree<=500